# src/main/resources/openapi/api-spec.yaml
openapi: 3.0.3
info:
  title: IA MSA DM Agent API
  description: |
    API REST para integración con modelos de IA mediante Spring AI.
    Proporciona endpoints para interactuar con modelos de lenguaje (LLM) como GPT-4, Azure OpenAI y Ollama.
  version: 1.0.0
  contact:
    name: IA Project Team
    email: support@iaproject.com
servers:
  - url: http://localhost:8080
    description: Servidor de desarrollo local
  - url: https://api.iaproject.com
    description: Servidor de producción
...
paths:
...
  /api/v1/events/{eventSlug}/rsvp:
    post:
      tags:
        - RSVP
      summary: Registrar confirmación de asistencia
      description: Crea o actualiza el RSVP de un invitado
      operationId: createRSVP
      parameters:
        - name: eventSlug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RSVPRequest'
      responses:
        '201':
          description: RSVP registrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RSVPResponse'
  /api/v1/events/{eventSlug}/rsvp/{userId}:
    get:
      tags:
        - RSVP
      summary: Consultar RSVP de un usuario
      description: Obtiene el RSVP de un usuario específico
      operationId: getRSVP
      parameters:
        - name: eventSlug
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: RSVP del usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RSVPResponse'
        '404':
          description: RSVP no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    put:
      tags:
        - RSVP
      summary: Actualizar RSVP
      description: Actualiza la confirmación de asistencia de un usuario
      operationId: updateRSVP
      parameters:
        - name: eventSlug
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RSVPRequest'
      responses:
        '200':
          description: RSVP actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RSVPResponse'
  /api/v1/events/{eventSlug}/attendees:
    get:
      tags:
        - RSVP
      summary: Listar confirmados
      description: Lista de invitados que confirmaron asistencia (status=YES)
      operationId: getAttendees
      parameters:
        - name: eventSlug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Lista de confirmados
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RSVPResponse'
  /api/v1/events/{eventSlug}/rsvps:
    get:
      tags:
        - RSVP
      summary: Listar todos los RSVPs
      description: Lista completa de RSVPs con resumen estadístico (organizador)
      operationId: getAllRSVPs
      parameters:
        - name: eventSlug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Lista de RSVPs con resumen
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RSVPListResponse'

  # ============================================================================
  # REGALOS
  # ============================================================================
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - timestamp
        - status
      properties:
        timestamp:
          type: string
          format: date-time
          description: Marca de tiempo del error
          example: "2025-12-17T10:30:00"
        status:
          type: integer
          format: int32
          description: Código de estado HTTP
          example: 400
        message:
          type: string
          description: Mensaje descriptivo del error
          example: "Error interno del servidor"
        details:
          type: string
          description: Detalles adicionales del error
          example: "Error al procesar el mensaje con el modelo de IA"
        errors:
          type: object
          description: Mapa de errores de validación por campo
          additionalProperties:
            type: string
          example:
            message: "El mensaje no puede estar vacío"
...
    RSVPRequest:
      type: object
      required:
        - userId
        - status
      properties:
        userId:
          type: string
        guestName:
          type: string
        guestEmail:
          type: string
        guestPhone:
          type: string
        status:
          type: string
          enum: [CONFIRMED, DECLINED, PENDING]
        guestsCount:
          type: integer
        notes:
          type: string
    RSVPResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        eventId:
          type: integer
          format: int64
        userId:
          type: string
        guestName:
          type: string
        guestEmail:
          type: string
        guestPhone:
          type: string
        status:
          type: string
          enum: [CONFIRMED, DECLINED, PENDING]
        guestsCount:
          type: integer
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    RSVPListResponse:
      type: object
      properties:
        rsvps:
          type: array
          items:
            $ref: '#/components/schemas/RSVPResponse'
        summary:
          $ref: '#/components/schemas/RSVPSummary'
    RSVPSummary:
      type: object
      properties:
        totalYes:
          type: integer
          format: int64
        totalNo:
          type: integer
          format: int64
        totalPending:
          type: integer
          format: int64
        totalGuests:
          type: integer
          format: int64

    # ============================================================================
    # SCHEMAS DE REGALOS
    # ============================================================================
...